Policy fixes
There's one final syntax error in policies/cost-control.rego line 93: "var _ is unsafe".

Please find any usage of `[_]` in the cost-control.rego file and replace it with `[i]` or another proper variable name. The underscore `_` is not safe in this context with the older OPA syntax.

For example:
- Change `input.resource_changes[_]` to `input.resource_changes[i]`
- Change `resource.change.actions[_]` to `resource.change.actions[j]`

This should resolve the remaining policy loading error.